Закрытие проекта

+++ src/main/resources/js/components/patternTable/tableArchive.vue
@@ -148,7 +148,7 @@
                                     icon="el-icon-view"/>
                             <span v-if="table.isArchive">
                             <el-button
-                                    @click="deArchiveOneTable(table.id)"
+                                    @click="deArchiveOneTable(table.id, table.sourceId)"
                                     style="margin-bottom: 10px; background-color: #1ab394; border-color: #1ab394 "
                                     type="primary"
                                     size="mini"
@@ -156,7 +156,7 @@
                         </span>
                             <span v-else>
                             <el-button
-                                    @click="deleteOneTable(table.id)"
+                                    @click="deleteOneTable(table.id, table.sourceId)"
                                     style="margin-bottom: 10px; background-color: #1ab394; border-color: #1ab394 "
                                     type="primary"
                                     size="mini"
@@ -350,10 +350,16 @@
                 });
             },
 
-            deleteOneTable(id) {
-                this.deleteTable(id);
+            deleteOneTable(id, sourceId) {
+                console.log(this.patternTableData);
+                AXIOS.get("user/isChangeSource/" + getToken() + "/" + sourceId).then(response=>{
+                    if(response.data) {
+                        this.deleteTable(id);
+                    } else {
+                        this.notify('Ошибка', 'У вас недостаточно прав', 'error');
+                    }
+                });
             },
-
             updatePage(){
                 if(this.patternTable.dateCreation !== null && this.patternTable.dateCreation !== "") {
                     this.patternTable.dateCreation1 = this.patternTable.dateCreation[0];
@@ -425,8 +431,16 @@
                 });
             },
 
-            deArchiveOneTable(id){
-                this.deArchiveTable(id);
+            deArchiveOneTable(id, sourceId){
+                console.log(this.patternTableData);
+                AXIOS.get("user/isChangeSource/" + getToken() + "/" + sourceId).then(response=>{
+                    if(response.data) {
+                        this.deArchiveTable(id);
+                    } else {
+                        this.notify('Ошибка', 'У вас недостаточно прав', 'error');
+                    }
+                });
+                console.log(sourceId);
             },
 
             onCurrentChange(value) {
